Refactoring candidat slicer
